revert all but time_t

+++ sys/arch/sh3/include/ansi.h
@@ -1,4 +1,4 @@
-/*	$NetBSD: ansi.h,v 1.12.18.2 2008/03/30 21:02:00 christos Exp $	*/
+/*	$NetBSD: ansi.h,v 1.12.18.3 2009/01/04 19:23:36 christos Exp $	*/
 
 /*-
  * Copyright (c) 1990, 1993
@@ -46,9 +46,9 @@
  *	#endif
  */
 #define	_BSD_CLOCK_T_		unsigned long	/* clock() */
-#define	_BSD_PTRDIFF_T_		long		/* ptr1 - ptr2 */
-#define	_BSD_SIZE_T_		unsigned long	/* sizeof() */
-#define	_BSD_SSIZE_T_		long		/* byte count or error */
+#define	_BSD_PTRDIFF_T_		int		/* ptr1 - ptr2 */
+#define	_BSD_SIZE_T_		unsigned int	/* sizeof() */
+#define	_BSD_SSIZE_T_		int		/* byte count or error */
 #define	_BSD_TIME_T_		__int64_t	/* time() */
 #if __GNUC_PREREQ__(2, 96)
 #define	_BSD_VA_LIST_		__builtin_va_list /* GCC built-in type */
